Feature/fix db

See merge request daigo26/startdash!17

+++ db/migrate/20180928230327_add_columns_to_posts.rb
@@ -1,6 +1,6 @@
 class AddColumnsToPosts < ActiveRecord::Migration[5.2]
   def change
-    add_column :posts, :rating, :double
+    add_column :posts, :rating, :float
     add_column :posts, :rating_count, :integer
     add_column :posts, :review_count, :integer
     add_column :posts, :new_feedback_count, :integer
